Pipelines table schedule now shows military time
